Blur, edge, sharp added + fixes

+++ color.cpp
@@ -2,4 +2,4 @@
 
 Color::Color() : r(0), g(0), b(0) {}
 
-Color::Color(float r, float g, float b) : r(r), g(g), b(b) {}
+Color::Color(int r, int g, int b) : r(r), g(g), b(b) {}
