+++ color.h
@@ -1,8 +1,8 @@
 #pragma once
 
 struct Color {
-    float r, g, b;
+    int r, g, b;
 
     Color();
-    Color(float r, float g, float b);
+    Color(int r, int g, int b);
 };
